chore(backend): clear the quality-gate backlog and swap python-jose for PyJWT
The Stop-hook gate (ruff + mypy + pytest) was failing on 107 ruff and 50 mypy
errors that predate this work, which blocked any backend edit from ending a
turn. Clearing it first so the rest of the refactor can land incrementally.

- ruff: format + autofix repo-wide (E501, UP0xx, F401, I001), plus two B904
  `raise ... from None` in the token-parse paths. Configure
  extend-immutable-calls for fastapi.Query/Depends/Path/Body/Header, since
  calling them in a parameter default is the framework idiom rather than the
  mutable-default bug B008 exists to catch. Ignore UP042: rewriting
  `class X(str, Enum)` to StrEnum changes how SQLAlchemy's
  Enum(native_enum=False) resolves stored values, so it waits for the schema pass.

- exports.py: 34 of the 50 mypy errors came from reusing one `stmt` local
  across seven independent entity branches. Split into per-export builder
  functions behind an ExportBuilder mapping, returning rows as an Iterable so
  streaming can be dropped in later without touching the call site. Removes the
  unreachable trailing 404 — the ExportKey Literal makes FastAPI return 422
  first (audit r).

- Numeric columns on FeedRecord/HealthRecord were annotated Mapped[float] but
  SQLAlchemy returns Decimal, so every assignment from a Decimal-typed schema
  failed to typecheck. Corrected the annotations rather than casting at the
  call sites.

- python-jose 3.3.0 carries known algorithm-confusion and DoS advisories and
  ships no type information. Replaced with PyJWT 2.13.0, which resolves audit p
  and 7 mypy errors together.

- _validate_cycle_dates took a parameter named `status` that shadowed
  fastapi.status, so both of its raise paths threw AttributeError instead of
  returning 400. Renamed to new_status (audit c) — mypy flagged this once the
  file was clean enough to check.

Also: delete three stray untracked SQLite files, rewrite the corrupt 28-byte
UTF-16 README as UTF-8, and add the missing .dockerignore (builds were shipping
~300 MB of node_modules and .venv into the daemon context) and .env.example.

ruff 107 -> 0, mypy 50 -> 0.

+++ backend/app/models/feed_record.py
@@ -1,4 +1,5 @@
 from datetime import date, datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import Date, DateTime, ForeignKey, Numeric, String, func
@@ -8,7 +9,6 @@ from app.db.base import Base
 
 if TYPE_CHECKING:
     from app.models.hog import Hog
-    from app.models.user import User
 
 
 class FeedRecord(Base):
@@ -16,8 +16,8 @@ class FeedRecord(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     hog_id: Mapped[int] = mapped_column(ForeignKey("hogs.id"), nullable=False, index=True)
-    feed_amount: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
-    feed_cost: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
+    feed_amount: Mapped[Decimal] = mapped_column(Numeric(12, 4), nullable=False)
+    feed_cost: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
     currency_code: Mapped[str] = mapped_column(String(3), nullable=False, default="GBP")
     record_date: Mapped[date] = mapped_column(Date, nullable=False, index=True)
     created_by_user_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"), nullable=True)
