+++ backend/app/models/health_record.py
@@ -1,14 +1,14 @@
 from datetime import date, datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Date, DateTime, ForeignKey, Numeric, String, Text, func
+from sqlalchemy import Date, DateTime, ForeignKey, Numeric, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
 
 if TYPE_CHECKING:
     from app.models.hog import Hog
-    from app.models.user import User
 
 
 class HealthRecord(Base):
@@ -16,8 +16,8 @@ class HealthRecord(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     hog_id: Mapped[int] = mapped_column(ForeignKey("hogs.id"), nullable=False, index=True)
-    weight: Mapped[float] = mapped_column(Numeric(10, 3), nullable=False)
-    temperature: Mapped[float | None] = mapped_column(Numeric(5, 2), nullable=True)
+    weight: Mapped[Decimal] = mapped_column(Numeric(10, 3), nullable=False)
+    temperature: Mapped[Decimal | None] = mapped_column(Numeric(5, 2), nullable=True)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
     record_date: Mapped[date] = mapped_column(Date, nullable=False, index=True)
     created_by_user_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"), nullable=True)
